some more cleaning up
