--- public/css/home.css
@@ -1,161 +1,0 @@
-body, html {
-    height: 100%;
-  }
-  .bg {
-   
-    background-image: url("../images/architecture-buildings-city-219692.jpg");
-  
-    /* Full height */
-    height: 100%;
-  
-    /* Center and scale the image nicely */
-    background-position: center;
-    background-repeat: no-repeat;
-    background-size: cover;
-  }
-  .sondertitle {
-    color:white;
-    font-size: 150px;
-    text-align: center;
-    font-family: 'Caveat', cursive;
-  }
-  
-  .logincontainer{
-    /*border:1px solid white;*/
-    width: 600px;
-    height: 350px;
-    position: absolute;
-    top:50%;
-    left:50%;
-    transform: translate(-50%, -50%);
-    display: inline-flex; 
-  }
-  .backbox{  
-    background-color: #404040;
-    border: none;
-    width: 100%;
-    height: 80%;
-    position: absolute;
-    transform: translate(0,-50%);
-    top:50%;
-    display: inline-flex;
-  }
-  
-  .frontbox{
-    background-color: white;
-    border-radius: 20px;
-    border: none;
-    height: 100%;
-    width: 50%;
-    z-index: 10;
-    position: absolute;
-    right:0;
-    margin-right: 3%;
-    margin-left: 3%;
-    transition: right .8s ease-in-out;
-  }
-  
-  .moving{
-    right:45%;
-  }
-  
-  .loginMsg, .signupMsg{
-    width: 50%;
-    height: 100%;
-    font-size: 15px;
-    box-sizing: border-box;
-  }
-  
-  .loginMsg .title,
-  .signupMsg .title{
-    font-weight: 300;
-    font-size: 23px;
-  }
-  
-  .loginMsg p,
-  .signupMsg p {
-    font-weight: 100;
-  }
-  
-  .textcontent{
-    color:white;
-    margin-top:65px;
-    margin-left: 12%;
-  }
-  
-  .loginMsg button,
-  .signupMsg button {
-    background-color: #404040;
-    border: 2px solid white;
-    border-radius: 10px;
-    color:white;
-    font-size: 12px;
-    box-sizing: content-box;
-    font-weight: 300;
-    padding:10px;
-    margin-top: 20px;
-  }
-  
-  /* front box content*/
-  .login, .signup{
-    padding: 10px;
-    text-align: center;
-  }
-  
-  .login h2,
-  .signup h2 {
-    color: #c2bb7ce1;
-    font-size:22px;
-  }
-  
-  .inputbox{
-    margin-top:30px; 
-  }
-  .login input, 
-  .signup input {
-    display: block;
-    width: 100%;
-    height: 35px;
-    background-color: #f2f2f2;
-    border: none;
-    margin-bottom:20px;
-    font-size: 12px;
-  }
-  
-  
-  .login button,
-  .signup button{
-    background-color: #c0b574c0;
-    border: none;
-    color:white;
-    font-size: 12px;
-    font-weight: 300;
-    box-sizing: content-box;
-    padding:10px;
-    border-radius: 10px;
-    width: 60px;
-    position: absolute;
-    right:30px;
-    bottom: 30px;
-    cursor: pointer;
-  }
-  
-  /* Fade In & Out*/
-  .login p {
-    cursor: pointer;
-    color:#404040;
-    font-size:15px;
-  }
-  
-  .loginMsg, .signupMsg{
-    /*opacity: 1;*/
-    transition: opacity .8s ease-in-out;
-  }
-  
-  .visibility{
-    opacity: 0;
-  }
-  
-  .hide{
-    display: none;
-  }
