--- routes/api/users.js
@@ -1,12 +1,0 @@
-var db = require("../../models");
-var express = require('express');
-var router  = express.Router();
-const passport = require('../../utils/middleware/passport-local');
-const userController = require('../../controllers/userController');
-
-
-
-
-  
-
-module.exports = router;
