+++ views/home.handlebars
@@ -34,16 +34,16 @@
             <div class="navbar-end">
               <span class="navbar-item">
                 <span class="navbar-item">
-                <a href="./view/home.handlebars#about" class="button is-success is-inverted"> About Us </a>
+                <a href="./view/home.handlebars#about" class="button has-text-warning-dark"> About Us </a>
                 </span>
                 <span class="navbar-item">
-                  <a href="/trips" class="button is-success is-inverted" class= "hide"> Trips </a>
+                  <a href="/trips" class="button has-text-warning-dark" class= "hide"> Trips </a>
                 </span>
                 <span class="navbar-item">
-                    <a href="./view/home.handlebars#search" class="button is-success is-inverted"> Signup/Login </a>
+                    <a href="./view/home.handlebars#search" class="button has-text-warning-dark"> Signup/Login </a>
                 </span>
                 <span class="navbar-item">
-                    <a href="/logout" class="button is-success is-inverted"> Logout </a>
+                    <a href="/logout" class="button has-text-warning-dark"> Logout </a>
                 </span>
             </div>
           </div>
